refactor: 뷰와 리듀서 분리
ListItemView / ListItemFeature

SearchBar / SearchFeature

파일 분리
