+++ CleanArea/CleanArea/iOS/Main/Components/SearchBar.swift
@@ -8,32 +8,6 @@ import ComposableArchitecture
 
 import SwiftUI
 
-@Reducer
-struct SearchFeature {
-    @ObservableState
-    struct State: Equatable {
-        var text: String = ""
-    }
-    
-    enum Action {
-        case clearTextField
-        case setText(String)
-    }
-    
-    var body: some ReducerOf<Self> {
-        Reduce { state, action in
-            switch action {
-            case .clearTextField:
-                state.text = ""
-                return .none
-            case let .setText(text):
-                state.text = text
-                return .none
-            }
-        }
-    }
-}
-
 struct SearchBar: View {
     @Bindable var store: StoreOf<SearchFeature>
